Data Anotations and DTO in backend developed

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,14 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using ProEventos.Persistence;
-using ProEventos.Domain;
-using ProEventos.Persistence.Contextos;
 using ProEventos.Application.Contratos;
 using Microsoft.AspNetCore.Http;
+using Back.src.ProEventos.Aplication.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -16,8 +12,7 @@ namespace ProEventos.API.Controllers
     [Route("api/[controller]")]
     public class EventosController : ControllerBase
     {
-
-
+        
         private readonly IEventoService _eventoService;
 
 
@@ -34,11 +29,11 @@ namespace ProEventos.API.Controllers
             try 
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if(eventos ==  null)
-                {
-                      return NotFound("Nenhum evento encontrado.");  
-                }
 
+                if(eventos ==  null) return NoContent();  
+                
+                
+                //usando dto quem está consumindo nao tem noção dos outros campos, ele nao tem acesso direto ao meu domínio
                 return Ok(eventos);
             }
             catch(Exception ex)
@@ -55,10 +50,7 @@ namespace ProEventos.API.Controllers
             try 
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id,true);
-                if(evento ==  null)
-                {
-                      return NotFound("Evento por Id não encontrado");  
-                }
+                if(evento ==  null) return NoContent();
 
                 return Ok(evento);
             }
@@ -75,10 +67,7 @@ namespace ProEventos.API.Controllers
             try 
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema,true);
-                if(evento ==  null)
-                {
-                      return NotFound("Eventos por tema não encontrado.");  
-                }
+                if(evento ==  null) return NoContent();
 
                 return Ok(evento);
             }
@@ -92,14 +81,13 @@ namespace ProEventos.API.Controllers
 
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try 
             {
                 var evento = await _eventoService.AddEventos(model);
-                if(evento ==  null) return BadRequest("Erro ao tentar adicionar eventos.");  
+                if(evento ==  null) return NoContent(); 
                 
-
                 return Ok(evento);
             }
             catch(Exception ex)
@@ -110,13 +98,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try 
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if(evento ==  null) return BadRequest("Erro ao tentar adicionar eventos.");  
-                
+                if(evento ==  null) return NoContent();   
 
                 return Ok(evento);
             }
@@ -132,9 +119,12 @@ namespace ProEventos.API.Controllers
         {
             try 
             {
-              return await _eventoService.DeleteEvento(id) ?
+                var evento = await _eventoService.GetEventoByIdAsync(id,true);
+                if(evento ==  null) return NoContent();
+
+                return await _eventoService.DeleteEvento(id) ?
                      Ok("Deletado") : 
-                     BadRequest("Evento não deletado");
+                     throw new Exception("Ocorreu um problema não específico ao tentar deletar o Evento.");
 
             }
             catch(Exception ex)
@@ -142,6 +132,7 @@ namespace ProEventos.API.Controllers
                 return this.StatusCode(StatusCodes.Status500InternalServerError,
                 $"Erro ao tentar deletar o evento. Erro: {ex.Message}");
             }
+            
         }
 
     }
